Bug fixes.
Fix list/grid mode switch not working on BookmarksController.
Fix crash when clicking "Move bookmark to a group" which launched a new screen and also modified the toolbar on the previous screen simultaneously so one of the sanity checks was triggered by it. In the end, the check was removed because that's a valid case.