Implement custom logic to show/hide keyboard from the toolbar.
Keep a set to toolbars that are currently requesting the keyboard to be visible. On each request to show/hide a keyboard add/remove the toolbar from the set. When empty close the keyboard. Use 250ms debounce time before hiding/showing the keyboard so that it doesn't jerk around when show/hide requests happen too often.