Fix composite catalog sorting issues.
When catalog sorting is set to anything other than bump - load the whole composite catalog because otherwise after incrementally loading each page the actual orders of threads would be very hard to guess (due to sorting). Since none of the imageboards support composite catalogs, the only way to fix it is to load composite catalogs fully and sort them afterwards.